requirements.txt
datasets==3.1.0
einops==0.8.0
ftfy==6.3.1
geom-median==0.1.0
gradio==5.8.0
jaxtyping==0.2.36
matplotlib==3.9.3
nbformat
numpy==1.26.4
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
regex==2024.11.6
Requests==2.32.3
scipy==1.14.1
setuptools==75.1.0
torchvision==0.17.2
tqdm==4.67.1
transformer_lens==2.9.1
transformers==4.46.3
wandb==0.19.0
yacs==0.1.8
